pyproject.toml
[tool.poetry]
name = "csv_reader"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.32"
pandas = "^2.2.2"
apscheduler = "^3.10.4"
gdown = "^4.7.1"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"
uvicorn = "^0.30.6"
pytest = "^8.3.2"
pytest-sqlalchemy = "^0.2.1"

[tool.poetry.scripts]
lint = "scripts.lint:main"
format = "scripts.black_format:main"
tests = "scripts.run_tests:main"
app = "main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

